# code/backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/games/{id}:
    get:
      tags:
      - game-rest-controller
      summary: Get a game by its id
      operationId: gameDetails
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Game not found
          content: {}
        "200":
          description: Found the game
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDetails'
        "403":
          description: forbiden o dont have permissions
    put:
      tags:
      - game-rest-controller
      summary: Edit a game
      operationId: editGame
      parameters:
      - name: game
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Game'
      - name: imageField
        in: query
        required: true
        schema:
          type: string
          format: binary
      - name: imageFields
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
            format: binary
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Game not found
          content: {}
        "200":
          description: Game edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
    delete:
      tags:
      - game-rest-controller
      summary: Delete a game
      operationId: deleteGame
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Game not found
          content: {}
        "200":
          description: Game deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
  /api/games/:
    get:
      tags:
      - game-rest-controller
      summary: Get a page of games
      operationId: getGames
      responses:
        "404":
          description: Games not found
          content: {}
        "200":
          description: Found the games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
    post:
      tags:
      - game-rest-controller
      summary: Creates a new game
      operationId: newGameProcess
      parameters:
      - name: game
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Game'
      - name: imageField
        in: query
        required: true
        schema:
          type: string
          format: binary
      - name: imageFields
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
            format: binary
      responses:
        "404":
          description: Not Found
        "201":
          description: Game created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
  /api/users/{id}/imageProfile:
    get:
      tags:
      - user-rest-controller
      summary: Get the image profile of the user
      operationId: downloadImageProfile
      parameters:
      - name: id
        in: path
        description: id of the current user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: image profile not found
          content: {}
        "200":
          description: Found the image profile
          content:
            image/png: {}
        "403":
          description: forbiden o dont have permissions
...
  /api/games/{id}/graphics:
    get:
      tags:
      - game-rest-controller
      summary: Get the graphics of a game
      operationId: getGraphics
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Game not found
          content: {}
        "200":
          description: Found the praphics of the game
          content:
            application/json:
              schema:
                type: array
                format: int32
  /api/games/{id}/gameplayImage/{index}:
    get:
      tags:
      - game-rest-controller
      summary: Get the gameplay images of a game
      operationId: downloadGameplayImages
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      - name: index
        in: path
        description: index of the gameplay image
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Gameplay images not found
          content: {}
        "200":
          description: Found the gameplay images
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /api/games/{id}/coverImage:
    get:
      tags:
      - game-rest-controller
      summary: Get the cover image of a game
      operationId: downloadImageProfile_1
      parameters:
      - name: id
        in: path
        description: id of the game
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Cover not found
          content: {}
        "200":
          description: Found the cover image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /api/games/search:
    get:
      tags:
      - game-rest-controller
      summary: Get a page of games by name or category
      operationId: search
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: category
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Games not found
          content: {}
        "200":
          description: Found the games by name or category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /api/games/moreIndexGames/{page}:
    get:
      tags:
      - game-rest-controller
      summary: Get more games
      operationId: getMoreIndexGames
      parameters:
      - name: page
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Games not found
          content: {}
        "200":
          description: Found the games
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
  /api/games/moreFoundGames/{page}:
    get:
      tags:
      - game-rest-controller
      summary: Get more games by name or category
      operationId: getMoreFoundGames
      parameters:
      - name: page
        in: path
        description: page of games
        required: true
        schema:
          type: integer
          format: int32
      - name: category
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Games not found
          content: {}
        "200":
          description: Found the games by name or category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        mail:
          type: string
        name:
          type: string
        lastName:
          type: string
        aboutMe:
          type: string
        billingInformation:
          type: string
        roles:
          type: array
          items:
            type: string
        profilePircture:
          type: string
        totalPrice:
          type: number
          format: float
    Game:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
          format: float
        category:
          type: string
        os:
          type: string
        processor:
          type: string
        memory:
          type: string
        graphics:
          type: string
        directX:
          type: string
        network:
          type: string
        hardDrive:
          type: string
        soundCard:
          type: string
        deleted:
          type: boolean
        totalRating:
          type: integer
          format: int32
        starDistribution:
          type: array
          items:
            type: integer
            format: int32
        gameplayImages:
          type: array
          items:
            type: string
        titleImage:
          type: string
        reviewCount:
          type: integer
          format: int32
        starDistributionInt:
          type: array
          items:
            type: integer
            format: int32
...
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user:
          $ref: '#/components/schemas/User'
        game:
          $ref: '#/components/schemas/Game'
        date:
          type: string
          format: date
        rating:
          type: integer
          format: int32
        comment:
          type: string
...
    GameDetails:
      type: object
      properties:
        game:
          $ref: '#/components/schemas/Game'
        reviews:
          $ref: '#/components/schemas/PageReview'
    PageReview:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        pageable:
          $ref: '#/components/schemas/Pageable'
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean